daily_retro/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.118.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
